respond to my own comments
