+++ components/src/dynamo/global_planner/tests/unit/test_kubernetes_capacity_manager.py
@@ -7,9 +7,9 @@ from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 
-from dynamo.global_planner.capacity_manager import (
+from dynamo.global_planner.capacity_manager import PoolSpec
+from dynamo.global_planner.kubernetes_capacity_manager import (
     KubernetesCapacityManager,
-    PoolSpec,
 )
 from dynamo.planner import SubComponentType, TargetReplica
 
@@ -84,9 +84,11 @@ def test_managed_deployment_names_mismatched_prefix():
 def test_discover_explicit_mode():
     cm = KubernetesCapacityManager("default")
     with (
-        patch("dynamo.global_planner.capacity_manager.KubernetesAPI") as mock_kube_cls,
         patch(
-            "dynamo.global_planner.capacity_manager.KubernetesConnector"
+            "dynamo.global_planner.kubernetes_capacity_manager.KubernetesAPI"
+        ) as mock_kube_cls,
+        patch(
+            "dynamo.global_planner.kubernetes_capacity_manager.KubernetesConnector"
         ) as mock_connector_cls,
     ):
         mock_kube = MagicMock()
@@ -110,9 +112,11 @@ def test_discover_explicit_mode():
 def test_discover_implicit_mode():
     cm = KubernetesCapacityManager("default")
     with (
-        patch("dynamo.global_planner.capacity_manager.KubernetesAPI") as mock_kube_cls,
         patch(
-            "dynamo.global_planner.capacity_manager.KubernetesConnector"
+            "dynamo.global_planner.kubernetes_capacity_manager.KubernetesAPI"
+        ) as mock_kube_cls,
+        patch(
+            "dynamo.global_planner.kubernetes_capacity_manager.KubernetesConnector"
         ) as mock_connector_cls,
     ):
         mock_kube = MagicMock()
@@ -134,7 +138,7 @@ def test_discover_implicit_mode():
 def test_discover_tolerates_api_failure():
     cm = KubernetesCapacityManager("default")
     with patch(
-        "dynamo.global_planner.capacity_manager.KubernetesAPI",
+        "dynamo.global_planner.kubernetes_capacity_manager.KubernetesAPI",
         side_effect=RuntimeError("no cluster"),
     ):
         # Best-effort: must not raise, returns nothing discovered.
@@ -150,24 +154,24 @@ def test_discover_tolerates_api_failure():
 def test_ensure_participant_idempotent():
     cm = KubernetesCapacityManager("default")
     with patch(
-        "dynamo.global_planner.capacity_manager.KubernetesConnector"
+        "dynamo.global_planner.kubernetes_capacity_manager.KubernetesConnector"
     ) as mock_connector_cls:
         mock_connector_cls.return_value = MagicMock()
         cm.ensure_participant(
             "default/my-dgd",
             caller_name="app-ns",
-            cluster_name="default",
+            namespace="default",
             deployment_name="my-dgd",
         )
         cm.ensure_participant(
             "default/my-dgd",
             caller_name="app-ns",
-            cluster_name="default",
+            namespace="default",
             deployment_name="my-dgd",
         )
         assert mock_connector_cls.call_count == 1
-        assert cm.knows("default/my-dgd")
-        assert not cm.knows("default/other")
+        assert cm.participant_exists("default/my-dgd")
+        assert not cm.participant_exists("default/other")
 
 
 def test_observe_parses_pools():
@@ -209,11 +213,11 @@ def test_current_replicas():
 
 
 @pytest.mark.asyncio
-async def test_actuate_calls_connector():
+async def test_scale_calls_connector():
     cm = KubernetesCapacityManager("default")
     connector = _install_connector(cm, "default/my-dgd", _dgd_spec(1, 1))
     targets = [
         TargetReplica(sub_component_type=SubComponentType.PREFILL, desired_replicas=3)
     ]
-    await cm.actuate("default/my-dgd", targets, blocking=True)
+    await cm.scale("default/my-dgd", targets, blocking=True)
     connector.set_component_replicas.assert_awaited_once_with(targets, blocking=True)
